requirements.txt
certifi==2019.11.28
Click==7.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.4
numpy>=1.9.2
scipy>=0.15.1
matplotlib>=1.4.3
tpot>=0.11.0
cloudpickle>=0.6.0
pandas==0.24.0rc1
Flask==1.1.1
google-cloud-storage==1.26.0
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
scikit-learn
google.cloud


